Refactor : MessageSection과 ThreadSection에 중복되는 메서드 분리
messageInnerElement, onKeyDownMessageTextarea 메서드의 형태를 변경하여 ContentsSectionCommon 파일로 분리하였습니다.
